+++ lib/lottie/gui/__init__.py
@@ -1,5 +1,5 @@
-from . import import_export, tree_view, timeline_widget, console
+from . import import_export, tree_view, timeline_widget, console, search_widget
 
 __all__ = [
-    "import_export", "tree_view", "timeline_widget", "console"
+    "import_export", "tree_view", "timeline_widget", "console", "search_widget"
 ]