+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
  * SVG exporter improvements
  * Implemented Alpha/Luma mattes on SVG export
+ * Find/replace in lottie gui code editor
 
 ## 0.6.2 - 2020-04-30
  * Support for all importers / exporters in lottie GUI